Update docker-entrypoint.sh
The ifconfig command isn't interpolated  properly due to the missing end quote, and thus not appearing in stdout. Since the directions on cloudsuite.ch clearly instruct end-users to look for this ip for use on in the client container, this causes substantial confusion (if not potentially other errors in the script..).  Please fix this issue :)  Thanks!